+++ tests/variables/identifier.cpp
@@ -1,5 +1,5 @@
 
-#include <duct/variables.hpp>
+#include <duct/Variables.hpp>
 
 #include <stdio.h>
 #include <iostream>
@@ -10,7 +10,7 @@ using namespace duct;
 using namespace std;
 
 void printValues(Identifier const& iden) {
-	printf("Variable count: %lu\n", iden.getChildCount());
+	printf("Variable count: %lu\n", (unsigned long)iden.getChildCount());
 	icu::UnicodeString str;
 	for (VarList::const_iterator iter=iden.begin(); iter!=iden.end(); ++iter) {
 		ValueVariable* var=dynamic_cast<ValueVariable*>(*iter);
